requirements/testing.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    make upgrade
#
astroid==2.4.2            # via pylint
attrs==20.3.0             # via pytest
boto==2.49.0              # via -r requirements/base.in
certifi==2020.11.8        # via requests
chardet==3.0.4            # via requests
click==7.1.2              # via -r requirements/base.in
colorama==0.4.4           # via -r requirements/base.in
coverage==5.3             # via pytest-cov
dj-database-url==0.5.0    # via -r requirements/base.in
django-bootstrap3==14.2.0  # via -r requirements/base.in
git+https://github.com/behype-inc/django-datetime-widget.git@2a123692cf18614a60363cb1828dd29cec9ed986#egg=django-datetime-widget  # via -r requirements/base.in
django-dotenv==1.4.2      # via -r requirements/base.in
django-filter==2.4.0      # via -r requirements/base.in
django-model-utils==4.0.0  # via -r requirements/base.in
djangorestframework==3.12.2  # via -r requirements/base.in
gunicorn==20.0.4          # via -r requirements/production.in
idna==2.10                # via requests
iniconfig==1.1.1          # via pytest
isort==5.6.4              # via pylint
lazy-object-proxy==1.4.3  # via -r requirements/base.in, astroid
mccabe==0.6.1             # via pylint
mock==4.0.2               # via -r requirements/testing.in
model-mommy==2.0.0        # via -r requirements/testing.in
ordereddict==1.1          # via -r requirements/base.in
packaging==20.4           # via pytest
pluggy==0.13.1            # via pytest
psycopg2==2.8.6           # via -r requirements/production.in
py==1.9.0                 # via pytest
pygments==2.7.2           # via -r requirements/base.in
pylint==2.6.0             # via -r requirements/testing.in
pyparsing==2.4.7          # via packaging
pytest-cov==2.10.1        # via -r requirements/testing.in
pytest-django==4.1.0      # via -r requirements/testing.in
pytest==6.1.2             # via -r requirements/testing.in, pytest-cov, pytest-django
python-dateutil==2.8.1    # via -r requirements/base.in
python-dotenv==0.15.0     # via -r requirements/base.in
pytz==2020.4              # via -r requirements/base.in, django, django-datetime-widget
requests==2.24.0          # via -r requirements/base.in, slacker
six==1.15.0               # via astroid, packaging, python-dateutil
slacker==0.14.0           # via -r requirements/base.in
sqlparse==0.4.1           # via django
toml==0.10.2              # via pylint, pytest
urllib3==1.25.11          # via requests
urwid==2.1.2              # via -r requirements/base.in
whitenoise==5.2.0         # via -r requirements/production.in
wrapt==1.11.0             # via -c requirements/constraints.txt, -r requirements/base.in, astroid

# The following packages are considered to be unsafe in a requirements file:
# setuptools
